Use HUGE_VALF insted of -1 as the default value for "smallest" searches.
This patch rewrites "find the smallest value" "find the closest object"
loops to use the HUGE_VALF constant, changing the following construct:

    float closest = -1;
    Object *closestObject = 0;
    for (Object *object = firstObject(); object; object = nextObject()) {
        float dist = distanceToObject(object);
        if (closest == -1 || dist < closest)
            closest = dist, closestObject = object;
    }

by initializing the distance to the currently closest object to infinity
instead of a negative value.  (The constant HUGE_VALF is normally
defined in <math.h> as a single-precision infinity, but a fallback is
supplied in BBGE/Base.h for environments which are missing the constant.)
This allows the "closest == -1" test to be eliminated; depending on the
loop structure and execution architecture, removing that single
comparison can reduce execution time by several percent.

As a further optimization, loops which include a test for a maximum
value (such as "find the closest object within 100 units") have that
test merged into the distance test by initializing the closest-distance
variable to that limit instead of HUGE_VALF.

In addition, some loops which previously used integers to hold the
current and smallest distances have been changed to use floats, since
it is typically less expensive to directly compare two floating-point
values than to convert from floating-point to integer and then compare
two integers.

+++ Aquaria/Game.h
@@ -681,11 +681,11 @@ public:
 
 	bool collideCircleVsCircle(Entity *a, Entity *b);
 	Bone *collideSkeletalVsCircle(Entity *skeletal, Entity *circle);
-	Bone *collideSkeletalVsLine(Entity *skeletal, Vector start, Vector end, int radius);
-	bool collideCircleVsLine(Entity *ent, Vector start, Vector end, int radius);
-	bool collideCircleVsLineAngle(Entity *ent, float angle, int startLen, int endLen, int radius, Vector basePos);
-	Bone *collideSkeletalVsCircle(Entity *skeletal, Vector pos, int radius);
-	CollideData collideCircleWithAllEntities(Vector pos, int r, Entity *me=0, int spellType=0, bool checkAvatarFlag=false);//, bool checkSpellFlag=0, bool checkHitEntitiesFlag=1);
+	Bone *collideSkeletalVsLine(Entity *skeletal, Vector start, Vector end, float radius);
+	bool collideCircleVsLine(Entity *ent, Vector start, Vector end, float radius);
+	bool collideCircleVsLineAngle(Entity *ent, float angle, float startLen, float endLen, float radius, Vector basePos);
+	Bone *collideSkeletalVsCircle(Entity *skeletal, Vector pos, float radius);
+	CollideData collideCircleWithAllEntities(Vector pos, float r, Entity *me=0, int spellType=0, bool checkAvatarFlag=false);//, bool checkSpellFlag=0, bool checkHitEntitiesFlag=1);
 	void handleShotCollisions(Entity *e, bool hasShield=false);
 	void handleShotCollisionsSkeletal(Entity *e);
 	void handleShotCollisionsHair(Entity *e, int num = 0);
